Move run_*() functions into run module
This makes more conceptual sense than having them in the suite,
testcase and step modules because they all get called only when
`compass run` gets called.
